feat: move utils functions in common
