+++ libs/ledger-live-common/src/modularDrawer/utils/groupAccountsByAsset.ts
