+++ libs/ledger-live-common/src/modularDrawer/utils/parseToBigNumber.ts
